+++ TTauri/GUI/View.hpp
@@ -51,7 +51,7 @@ public:
     Device *device();
   
 
-    virtual off_t BackingPipelineRender(BackingPipeline::Vertex *vertices, off_t offset, size_t size);
+    virtual size_t BackingPipelineRender(BackingPipeline::Vertex *vertices, size_t offset, size_t size);
 };
 
 }}
